NetworkDynamics.jl

NetworkDynamics.jl

Overview

This package implements functions for defining and studying dynamics on networks. The key construction is a callable struct compatible with the DifferentialEquations.jl calling syntax.

nd = network_dynamics(nodes!, lines!, s_e, t_e, dim_n, dim_l; symbols_n=nothing, symbols_l=nothing)
nd(dx, x, p, t)

The first two parameters are the functions, or function arrays from which a network dynamics is constructed:

$$nodes![n](dx, x, [l]_s, [l]_t, p, t)$$ $$lines![e](dl, l, x_s, x_t, p, t)$$

The arrays dimn and diml encode the dimensionality of $x$ and $l$ variables. The arrays se and te encode the network structure by giving the source and target of each edge.

Optionally we can also specify an array of symbols per edge and node that allow convenience access to the different nodal and line dimensions.

Given edges $e$, and nodes $n$, as well as an orientation encoded by the source function $s(e)$ and the target function $t(e)$ this implements the system of ODEs:

$$\frac{dx_n}{dt} = dx_n$$

$$\frac{dl_e}{dt} = dl_e$$

with $dx$ and $dl$ calculated by

$$[l]_s = [l_e \text{ if } s(e) = n]$$

$$[l]_t = [l_e \text{ if } t(e) = n]$$

$$nodes![n](dx_n, x_n, [l]_s, [l]_t, p_n, t)$$

$$lines![e](dl_e, l_e, x_{s(e)}, x_{t(e)}, p_e, t)$$

Something that relaxes to a diffusive network would for example be implemented by

lines = (dl, l, x_1, x_2) -> dl .= 1000. * ((x_1 - x_2) - l)
nodes = (dx_n, x_n, l_s, l_t, p_n, t) -> dx_n .= f(x_n) - (sum(l_s) - sum(l_t))

The package also supplies a node and a line type that combine the the node function, the dimensionality and the symbols, as well as a constructor for the network dynamics from these types. Further Constructors are provided for LightGraphs:

nd=(nodes::Array{ODE_Node}, lines::Array{ODE_Line}, G::AbstractGraph)

Static lines

For static line relations we similarly have:

sl_nd = static_lines_network_dynamics(nodes!, lines!, s_e, t_e, ...)
sl_nd(dx, x, p, t)

With the convention for lines given by:

$$lines![e](l, x_s, x_t, p, t)$$

and otherwise as above. This implements the system of ODEs:

$$\frac{dx_n}{dt} = dx_n$$

with $dx$ calculated by

$$lines![e](l_e, x_{s(e)}, x_{t(e)}, p_e, t)$$

$$[l]_s = [l_e \text{ if } s(e) = n]$$

$$[l]_t = [l_e \text{ if } t(e) = n]$$

$$nodes![n](dx_n, x_n, [l]_s, [l]_t, p_n, t)$$

A diffusive network would be implemented by

lines = (l, x_1, x_2) -> l .= x_1 - x_2
nodes = (dx_n, x_n, l_s, l_t, p_n, t) -> dx_n .= f(x_n) - (sum(l_s) - sum(l_t))

The alternative constructor is given by:

nd=(nodes::Array{ODE_Node}, lines::Array{Static_Line}, G::AbstractGraph)

Network DAEs

Network SDEs

Network DDEs

Convenience functions for symbolic access to node variables

API

NetworkDynamics.diffusive_network_dynamics — Type.

diffusive_network_dynamics(L, nodes)

L: Matrix nodes: scalar function $x \rightarrow nodes(x)$

source

NetworkDynamics.diffusive_network_dynamics — Method.

(dnd::diffusive_network_dynamics)(dx, x, p, t)

Calling a struct of type diffusive network dynamics implements the ODE:

$\frac{dx_i}{dt} = nodes(x_i) + \sum_j L_{ij} x_j$

dnd = diffusive_network_dynamics(L, nodes)
dnd(dx, x, p, t)

source

NetworkDynamics.diffusive_network_dynamics — Method.

When called with a graph, the dynamics defaults to using the laplacian.

source

NetworkDynamics.network_dynamics — Type.

The key functions or function arrays are:

nodes: $nodes!(dx, x, [l]_s,[l]_t, p, t)$

lines: $lines!(dl, l, x_s, x_t, p, t)$

Given edges $e$, ans nodes $n$, as well as an orientation encoded by the source function $s(e)$ and the target function $t(e)$ this implements the system of ODEs:

$\frac{dx_n}{dt} = dx_n$

$\frac{dl_e}{dt} = dl_e$

with $dx$ and $dl$ calculated by

$[l]_s = [l_e \text{ if } s(e) = n]$

$[l]_t = [l_e \text{ if } t(e) = n]$

$nodes![n](dx_n, x_n, [l]_s, [l]_t, p_n, t)$

$lines![e](dl_e, l_e, x_{s(e)}, x_{t(e)}, p_e, t)$

Alternative design:

Something that relaxes to a diffusive network would for example be implemented by

lines = (dl, l, x_1, x_2) -> dl .= 1000. * ((x_1 - x_2) - l)
agg = (list_1, list_2) -> sum(list_1) - sum(list_2)

source

NetworkDynamics.network_dynamics_on_the_line — Type.

edge_list: list of edges nodes: scalar function $x \rightarrow nodes(x)$ lines: scalar function $x \rightarrow lines(x)$

Implements the ODE $\frac{dx_i}{dt} = nodes(x_i) + \sum_j A_{ij} lines(x_i - x_j)$

where $A$ is the adjacency matrix of $B$

source

NetworkDynamics.network_dynamics_on_lines — Method.

When called with a graph, the dynamics defaults to using the laplacian.

source