API

The following functions are designed for public use

Network Construction

NetworkDynamics.Network — Type

Network(nw::Network; kwargs...)

Rebuild the Network with same graph and vertex/edge functions but possibly different kwargs.

NetworkDynamics.dim — Method

dim(nw::Network)

Returns the number of dynamic states in the network, corresponts to the length of the flat state vector.

NetworkDynamics.pdim — Method

pdim(nw::Network)

Returns the number of parameters in the network, corresponts to the length of the flat parameter vector.

Vertex Functions

Missing docstring.

Missing docstring for StaticVertex. Check Documenter's build log for details.

NetworkDynamics.ODEVertex — Type

struct ODEVertex{F, OF, MM} <: NetworkDynamics.VertexFunction

Fields

name

f

sym

depth

psym

inputsym

obsf

obssym

symmetadata

metadata

mass_matrix

Edge Functions

Missing docstring.

Missing docstring for StaticEdge. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ODEEdge. Check Documenter's build log for details.

Coupling types

NetworkDynamics.Symmetric — Type

struct Symmetric <: Coupling end

Symmetric coupling type. The edge function f is evaluated once:

the dst vertex receives the first d values of the edge state,

the src vertex receives the same.

Here, d is the edge depth of the Network.

NetworkDynamics.AntiSymmetric — Type

struct AntiSymmetric <: Coupling end

AntiSymmetric coupling type. The edge function f is evaluated once:

the dst vertex receives the first d values of the edge state,

the src vertex receives (-1) of that.

Here, d is the edge depth of the Network.

NetworkDynamics.Directed — Type

struct Directed <: Coupling end

Directed coupling type. The edge function f is evaluated once:

the dst vertex receives the first d values of the edge state,

the src vertex receives nothing.

Here, d is the edge depth of the Network.

NetworkDynamics.Fiducial — Type

struct Fiducial <: Coupling end

Fiducial coupling type. The edge function f is evaluated once:

the dst vertex receives the 1:d values of the edge state,

the src vertex receives the d+1:2d values of the edge state.

Here, d is the edge depth of the Network.

Symbolic Indexing

Network Parameter Object

NetworkDynamics.NWParameter — Type

NWParameter(nw_or_nw_wraper, pflat)

Indexable wrapper for flat parameter array pflat. Needs Network or wrapper of Network, e.g. ODEProblem.

p = NWParameter(nw)
p.v[idx, :sym] # get parameter :sym of vertex idx
p.e[idx, :sym] # get parameter :sym of edge idx
p[s::Union{VPIndex, EPIndex}] # get parameter for specific index

Get flat array representation using pflat(p).

NetworkDynamics.NWParameter — Method

NWParameter(nw_or_nw_wraper;
            ptype=Vector{Float64}, pfill=filltype(ptype), default=true)

Creates "empty" NWParameter object for the Network/Wrapper nw with flat type ptype. The array will be prefilled with pfill (defaults to NaN).

If default=true the default parameter values attached to the network components will be loaded.

NetworkDynamics.NWParameter — Method

NWParameter(p::NWParameter; ptype=typeof(p.pflat))

Create NWParameter based on other parameter object, just convert type.

NetworkDynamics.NWParameter — Method

NWParameter(int::SciMLBase.DEIntegrator)

Create NWParameter object from integrator.

Network State Object

NetworkDynamics.NWState — Type

NWState(nw_or_nw_wrapper, uflat, [pflat], [t])

Indexable wrapper for flat state & parameter array. Needs Network or wrapper of Network, e.g. ODEProblem.

s = NWState(nw)
s.v[idx, :sym] # get state :sym of vertex idx
s.e[idx, :sym] # get state :sym of edge idx
s.p.v[idx, :sym] # get parameter :sym of vertex idx
s.p.e[idx, :sym] # get parameter :sym of edge idx
s[s::Union{VIndex, EIndex, EPIndex, VPIndex}] # get parameter for specific index

Get flat array representation using uflat(s) and pflat(s).

NetworkDynamics.NWState — Method

NWState(nw_or_nw_wrapper;
        utype=Vector{Float64}, ufill=filltype(utype),
        ptype=Vector{Float64}, pfill=filltype(ptype), default=true)

Creates "empty" NWState object for the Network/Wrapper nw with flat types utype & ptype. The arrays will be prefilled with ufill and pfill respectively (defaults to NaN).

If default=true the default state & parameter values attached to the network components will be loaded.

NetworkDynamics.NWState — Method

NWState(p::NWState; utype=typeof(uflat(s)), ptype=typeof(pflat(s)))

Create NWState based on other state object, just convert types.

NetworkDynamics.NWState — Method

NWState(p::NWParameter; utype=Vector{Float64}, ufill=filltype(utype), default=true)

Create NWState based on existing NWParameter object.

NetworkDynamics.NWState — Method

NWState(int::SciMLBase.DEIntegrator)

Create NWState object from integrator.

NetworkDynamics.uflat — Function

uflat(s::NWState)

Retrieve the wrapped flat array representation of the state.

NetworkDynamics.pflat — Function

pflat(p::NWParameter)
pflat(s::NWState)

Retrieve the wrapped flat array representation of the parameters.

Index generators

NetworkDynamics.vidxs — Function

vidxs([inpr], components=:, variables=:) :: Vector{VIndex}

Generate vector of symbolic indexes for vertices.

inpr: Only needed for name matching or : access. Can be Network, sol, prob, ...

components: Number/Vector, :, Symbol (name matches), String/Regex (name contains)

variables: Symbol/Number/Vector, :, String/Regex (all sym containing)

Examples:

vidxs(nw)                 # all vertex state indices
vidxs(1:2, :u)            # [VIndex(1, :u), VIndex(2, :u)]
vidxs(nw, :, [:u, :v])    # [VIndex(i, :u), VIndex(i, :v) for i in 1:nv(nw)]
vidxs(nw, "ODEVertex", :) # all symbols of all vertices with name containing "ODEVertex"

NetworkDynamics.eidxs — Function

vidxs([inpr], components=:, variables=:) :: Vector{EIndex}

Generate vector of symbolic indexes for edges.

inpr: Only needed for name matching or : access. Can be Network, sol, prob, ...

components: Number/Vector, :, Symbol (name matches), String/Regex (name contains)

variables: Symbol/Number/Vector, :, String/Regex (all sym containing)

Examples:

eidxs(nw)                # all edge state indices
eidxs(1:2, :u)           # [EIndex(1, :u), EIndex(2, :u)]
eidxs(nw, :, [:u, :v])   # [EIndex(i, :u), EIndex(i, :v) for i in 1:ne(nw)]
eidxs(nw, "FlowEdge", :) # all symbols of all edges with name containing "FlowEdge"

NetworkDynamics.vpidxs — Function

vpidxs([inpr], components=:, variables=:) :: Vector{VPIndex}

Generate vector of symbolic indexes for parameters. See vidxs for more information.

NetworkDynamics.epidxs — Function

epidxs([inpr], components=:, variables=:) :: Vector{EPIndex}

Generate vector of symbolic indexes for parameters. See eidxs for more information.

Symbolic Indices

NetworkDynamics.VIndex — Type

VIndex{C,S} <: SymbolicStateIndex{C,S}
idx = VIndex(comp, sub)

A symbolic index for a vertex state variable.

comp: the component index, either int or a collection of ints

sub: the subindex, either int, symbol or a collection of those.

VIndex(1, :P)      # vertex 1, variable :P
VIndex(1:5, 1)     # first state of vertices 1 to 5
VIndex(7, (:x,:y)) # states :x and :y of vertex 7

Can be used to index into objects supporting the SymbolicIndexingInterface, e.g. NWState, NWParameter or ODESolution.

See also: EIndex, VPIndex, EPIndex

NetworkDynamics.EIndex — Type

EIndex{C,S} <: SymbolicStateIndex{C,S}
idx = EIndex(comp, sub)

A symbolic index for an edge state variable.

comp: the component index, either int or a collection of ints

sub: the subindex, either int, symbol or a collection of those.

EIndex(1, :P)      # edge 1, variable :P
EIndex(1:5, 1)     # first state of edges 1 to 5
EIndex(7, (:x,:y)) # states :x and :y of edge 7

Can be used to index into objects supporting the SymbolicIndexingInterface, e.g. NWState, NWParameter or ODESolution.

See also: VIndex, VPIndex, EPIndex

NetworkDynamics.VPIndex — Type

VPIndex{C,S} <: SymbolicStateIndex{C,S}
idx = VPIndex(comp, sub)

A symbolic index into the parameter a vertex:

comp: the component index, either int or a collection of ints

sub: the subindex, either int, symbol or a collection of those.

Can be used to index into objects supporting the SymbolicIndexingInterface, e.g. NWParameter or ODEProblem.

See also: EPIndex, VIndex, EIndex

NetworkDynamics.EPIndex — Type

VEIndex{C,S} <: SymbolicStateIndex{C,S}
idx = VEIndex(comp, sub)

A symbolic index into the parameter a vertex:

comp: the component index, either int or a collection of ints

sub: the subindex, either int, symbol or a collection of those.

Can be used to index into objects supporting the SymbolicIndexingInterface, e.g. NWParameter or ODEProblem.

See also: VPIndex, VIndex, EIndex

Metadata API

Component Accessors

NetworkDynamics.dim — Method

dim(c::ComponentFunction)::Int

Retrieve the dimension of the component.

NetworkDynamics.sym — Function

sym(c::ComponentFunction)::Vector{Symbol}

Retrieve the symbols of the component.

NetworkDynamics.pdim — Method

pdim(c::ComponentFunction)::Int

Retrieve the parameter dimension of the component.

NetworkDynamics.psym — Function

psym(c::ComponentFunction)::Vector{Symbol}

Retrieve the parameter symbols of the component.

NetworkDynamics.obssym — Function

obssym(c::ComponentFunction)::Vector{Symbol}

Retrieve the observation symbols of the component.

NetworkDynamics.depth — Function

depth(c::ComponentFunction)::Int

Retrieve the depth of the component. The depth is the number of "outputs". For example, in a vertex function, a depth N means, that the connected edges receive the states 1:N.

NetworkDynamics.hasinputsym — Function

hasinputsym(c::ComponentFunction)

Checks if the optioan field inputsym is present in the component function.

NetworkDynamics.inputsym — Function

hasinputsym(c::VertexFunction)::Vector{Symbol}
hasinputsym(c::EdgeFunction)::NamedTuple with :src and :dst keys

Musst be called after hasinputsym returned true. Gives the inputsym vector(s). For vertex functions just a single vector, for edges it returns a named tuple (; src, dst) with two symbol vectors.

NetworkDynamics.coupling — Function

coupling(::EdgeFunction)
coupling(::Type{<:EdgeFunction})

Returns the coupling of the given EdgeFunction.

Component Metadata API

NetworkDynamics.metadata — Function

metadata(c::ComponentFunction)

Retrieve metadata object for the component.

See also metadata

NetworkDynamics.has_metadata — Method

has_metadata(c::ComponentFunction, key::Symbol)

Checks if metadata key is present for the component.

NetworkDynamics.get_metadata — Method

get_metadata(c::ComponentFunction, key::Symbol)

Retrieves the metadata key for the component.

NetworkDynamics.set_metadata! — Method

set_metadata!(c::ComponentFunction, key::Symbol, value)

Sets the metadata key for the component to value.

NetworkDynamics.has_graphelement — Function

has_graphelement(c)

Checks if the edge or vetex function function has the graphelement metadata.

NetworkDynamics.get_graphelement — Function

get_graphelement(c::EdgeFunction)::@NamedTuple{src::T, dst::T}
get_graphelement(c::VertexFunction)::Int

Retrieves the graphelement metadata for the component function. For edges this returns a named tupe (;src, dst) where both are either integers (vertex index) or symbols (vertex name).

NetworkDynamics.set_graphelement! — Function

set_graphelement!(c::EdgeFunction, src, dst)
set_graphelement!(c::VertexFunction, vidx)

Sets the graphelement metadata for the edge function. For edges this takes two arguments src and dst which are either integer (vertex index) or symbol (vertex name). For vertices it takes a single integer vidx.

Per-Symbol Metadata API

NetworkDynamics.symmetadata — Function

symmetadata(c::ComponentFunction)::Dict{Symbol,Dict{Symbol,Any}}

Retrieve the metadata dictionary for the symbols. Keys are the names of the symbols as they appear in sym, psym, obssym and inputsym.

See also symmetadata

NetworkDynamics.get_metadata — Method

get_metadata(c::ComponentFunction, sym::Symbol, key::Symbol)

Retrievs the metadata key for symbol sym.

NetworkDynamics.has_metadata — Method

has_metadata(c::ComponentFunction, sym::Symbol, key::Symbol)

Checks if symbol metadata key is present for symbol sym.

NetworkDynamics.set_metadata! — Method

set_metadata!(c::ComponentFunction, sym::Symbol, key::Symbol, value)
set_metadata!(c::ComponentFunction, sym::Symbol, pair)

Sets the metadata key for symbol sym to value.

NetworkDynamics.has_default — Function

has_default(c::ComponentFunction, sym::Symbol)

Checks if a default value is present for symbol sym.

NetworkDynamics.get_default — Function

get_default(c::ComponentFunction, sym::Symbol)

Returns the default value for symbol sym.

NetworkDynamics.set_default! — Function

set_default!(c::ComponentFunction, sym::Symbol, value)

Sets the default value for symbol sym to value.

NetworkDynamics.has_guess — Function

has_guess(c::ComponentFunction, sym::Symbol)

Checks if a guess value is present for symbol sym.

NetworkDynamics.get_guess — Function

get_guess(c::ComponentFunction, sym::Symbol)

Returns the guess value for symbol sym.

NetworkDynamics.set_guess! — Function

set_guess!(c::ComponentFunction, sym::Symbol, value)

Sets the guess value for symbol sym to value.

NetworkDynamics.has_init — Function

has_init(c::ComponentFunction, sym::Symbol)

Checks if a init value is present for symbol sym.

NetworkDynamics.get_init — Function

get_init(c::ComponentFunction, sym::Symbol)

Returns the init value for symbol sym.

NetworkDynamics.set_init! — Function

set_init!(c::ComponentFunction, sym::Symbol, value)

Sets the init value for symbol sym to value.

NetworkDynamics.has_bounds — Function

has_bounds(c::ComponentFunction, sym::Symbol)

Checks if a bounds value is present for symbol sym.

NetworkDynamics.get_bounds — Function

get_bounds(c::ComponentFunction, sym::Symbol)

Returns the bounds value for symbol sym.

NetworkDynamics.set_bounds! — Function

set_bounds!(c::ComponentFunction, sym::Symbol, value)

Sets the bounds value for symbol sym to value.

Initialization

NetworkDynamics.find_fixpoint — Function

find_fixpoint(nw::Network, [x0::NWState=NWState(nw)], [p::NWParameter=x0.p]; kwargs...)
find_fixpoint(nw::Network, x0::AbstractVector, p::AbstractVector; kwargs...)
find_fixpoint(nw::Network, x0::AbstractVector; kwargs...)

Convenience wrapper around SteadyStateProblem from SciML-ecosystem. Constructs and solves the steady state problem, returns found value wrapped as NWState.

NetworkDynamics.initialize_component! — Function

initialize_component!(cf::ComponentFunction; verbose=true, kwargs...)

Initialize a ComponentFunction by solving the corresponding NonlinearLeastSquaresProblem. During initialization, everyting which has a default value (see Metadata) is considered "fixed". All other variables are considered "free" and are solved for. The initial guess for each variable depends on the guess value in the Metadata.

The result is stored in the ComponentFunction itself. The values of the free variables are stored in the metadata field init.

The kwargs are passed to the nonlinear solver.

NetworkDynamics.init_residual — Function

init_residual(cf::T; t=NaN, recalc=false)

Calculates the residual |du| for the given component function for the values provided via default and init Metadata.

If recalc=false just return the residual determined in the actual initialization process.

See also initialize_component!.

Utils

Callbacks

NetworkDynamics.save_parameters! — Function

save_parameters!(integrator::SciMLBase.DEIntegrator)

Save the current parameter values in the integrator. Call this function inside callbacks if the parameter values have changed. This will store a timeseries of said parameters in the solution object, thus alowing us to recosntruct observables which depend on time-dependet variables.