Library

Public Interface

The following functions are designed for public use

NetworkDynamics.NDFunctions.ODEEdge

NetworkDynamics.NDFunctions.ODEVertex

NetworkDynamics.NDFunctions.StaticEdge

NetworkDynamics.NDFunctions.StaticVertex

NetworkDynamics.network_dynamics

NetworkDynamics.network_dynamics — Function

network_dynamics(vertices!, edges!, g)

Assembles the the dynamical equations of the network problem into an ODEFunction compatible with the DifferentialEquations.jl solvers. Takes as arguments an array of VertexFunctions vertices!, an array of EdgeFunctions edges! and a LightGraph.jl object g.

NetworkDynamics.NDFunctions.StaticVertex — Type

StaticVertex(f!, dim, sym)

Wrapper that ensures compatibility of a mutating function f! with the key constructor network_dynamics.

f! describes the local behaviour at a static node and has to respect the following calling syntax

f!(v, e_s, e_t, p, t) -> nothing

Here v, p and t are the usual arguments, while e_s and e_d are arrays containing the edges for which the described vertex is the source or the destination respectively.

dim is the number of independent variables in the vertex equations and sym is an array of symbols for these variables.

For more details see the documentation.

NetworkDynamics.NDFunctions.ODEVertex — Type

ODEVertex(f!, dim, mass_matrix, sym)

Wrapper that ensures compatibility of a mutating function f! with the key constructor network_dynamics.

f! describes the local behaviour at a dynamic node and has to respect the following calling syntax

f!(dv, v, e_s, e_t, p, t) -> nothing

Here dv, v, p and t are the usual ODE arguments, while e_s and e_d are arrays containing the edges for which the described vertex is the source or the destination respectively.

dim is the number of independent variables in the vertex equations and sym is an array of symbols for these variables. mass_matrix is an optional argument that defaults to the identity matrix I. If a mass matrix M is given the system M * dv = f! will be solved.

For more details see the documentation.

NetworkDynamics.NDFunctions.StaticEdge — Type

StaticEdge(f!, dim, sym)

Wrapper that ensures compatibility of a mutating function f! with the key constructor network_dynamics.

f! describes the local behaviour at a static edge and has to respect the following calling syntax

f!(e, v_s, v_t, p, t) -> nothing

Here e, p and t are the usual arguments, while v_s and v_d are arrays containing the vertices which are the source and destination of the described edge.

dim is the number of independent variables in the edge equations and sym is an array of symbols for these variables.

For more details see the documentation.

NetworkDynamics.NDFunctions.ODEEdge — Type

ODEEdge(f!, dim, mass_matrix, sym)

Wrapper that ensures compatibility of a mutating function f! with the key constructor network_dynamics.

f! describes the local behaviour at a dynamic edge and has to respect the following calling syntax

f!(de, e, v_s, v_t, p, t) -> nothing

Here de, e, p and t are the usual arguments, while v_s and v_d are arrays containing the vertices which are the source and destination of the described edge.

dim is the number of independent variables in the edge equations and sym is an array of symbols for these variables. For more details see the documentation. mass_matrix is an optional argument that defaults to the identity matrix I. If a mass matrix M is given the system M * de = f! will be solved.

For more details see the documentation.